code solution but pseudocode and notes need rev
Creating New Branch as the pseudocode needs to be updated to reflect the new code logic code is sound but notes and pseudocode do not reflect what is going on very much in some areas. New variables added some logic changed. Disciplining self to use proper code dev tools revision history and other things sometimes may be faster to spit out the code and breeze through the challenges but remember not to develop bad habits method to the madness.